requirements.txt
fastapi
uvicorn
requests
python-dotenv
pandas
cerebras-cloud-sdk
